Fixed the recursion problem.
With the %: force rule, the makefile was attempting to run rules named
Makefile and Makefile.include.  Removing the force prerequisit fixed this
issue.

The tradeoff is that make may avoid running a rule if a file with the name
already exists. (For example, clean may not work if a file named clean
exists).